first pass - seems ok. having some weird interactions with sublime/rspec, but code seems stable.
