+++ student.rb
@@ -1,15 +1,11 @@
-class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+require_relative 'person'
 
+class Student < Person
+
+ 
   def initialize(options = {})
+    super
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
-  end
-
-  def offer_high_five
-    "High five!"
   end
 
   def set_phase(num)
@@ -20,7 +16,7 @@ class Student
     else
       response = "Oooh, phase #{num}. I hope I'm ready!"
     end
-    @phase = num
+    super
     response
   end
